Window class, removed "renderer", sorting Data, local FS, Text resizing and storing

+++ src/Tile.h
@@ -22,8 +22,8 @@ public:
 	Tile();
 	virtual ~Tile();
 
-	short unsigned int tileID;
-	short unsigned int typeID;
+	Uint8 tileID;
+	Uint8 typeID;
 
 	static SDL_Rect InTileSetPos(const int tile_size, const int x, const int y);
 };
